NavBar component add % Layout apply
